Isolate users by org specified in JWT

+++ persistence/octo_store.go
@@ -11,9 +11,12 @@ import (
 type OctoStore struct{}
 
 func (OctoStore) FetchAll(ctx context.Context, database Database) ([]data.Octo, error) {
-	query := "select id, name from octo order by id"
+	query := `select o.id, o.name from octo o
+		join org on o.org_id = org.id
+		where org.name = $1
+		order by o.id`
 
-	rows, err := database.Query(ctx, query)
+	rows, err := database.Query(ctx, query, org(ctx))
 	if err != nil {
 		return nil, err
 	}
@@ -39,13 +42,15 @@ func (OctoStore) FetchAll(ctx context.Context, database Database) ([]data.Octo, 
 }
 
 func (OctoStore) FetchByName(ctx context.Context, database Database, name string, selectForUpdate bool) (data.Octo, error) {
-	query := "select id from octo where name = $1"
+	query := `select o.id from octo o
+		join org on o.org_id = org.id
+		where o.name = $1 and org.name = $2`
 	if selectForUpdate {
 		query += " for update"
 	}
 
 	var id int
-	err := database.QueryRow(ctx, query, name).Scan(&id)
+	err := database.QueryRow(ctx, query, name, org(ctx)).Scan(&id)
 	if err == sql.ErrNoRows {
 		return data.Octo{}, ErrNotFound
 	} else if err != nil {
@@ -59,13 +64,13 @@ func (OctoStore) FetchByName(ctx context.Context, database Database, name string
 }
 
 func (OctoStore) Create(ctx context.Context, database Database, octo data.Octo) (int, error) {
-	query := "insert into octo (name) values ($1) returning id"
-	return ExecInsert(ctx, database, query, octo.Name)
+	query := "insert into octo (name, org_id) values ($1, (select id from org where name = $2)) returning id"
+	return ExecInsert(ctx, database, query, octo.Name, org(ctx))
 }
 
 func (OctoStore) DeleteById(ctx context.Context, database Database, id int) error {
-	query := "delete from octo where id = $1"
-	rowsAffected, err := ExecDelete(ctx, database, query, id)
+	query := "delete from octo where id = $1 and org_id = (select id from org where name = $2)"
+	rowsAffected, err := ExecDelete(ctx, database, query, id, org(ctx))
 	if err != nil {
 		return err
 	}
